Brick checks empty or not. Title and Brief is required

+++ src/firebase/brick.ts
@@ -1,10 +1,24 @@
 var firestore = require('./firestore');
 var bricksRef = firestore.collection('bricks');
 
+/**
+ * Get Brick Reference
+ * @param brickId Brick Id
+ * @returns Brick Reference
+ */
 function brickRef(brickId) { return bricksRef.doc(brickId); }
 
 /**
- * Get all bricks
+ * Get Brick
+ * @param brickId Brick Id
+ * @returns Promise with Brick
+ */
+function getBrick(brickId) {
+  return brickRef(brickId).get().then(brick => brick.data());
+}
+
+/**
+ * Get all brick ids and titles
  * @returns {Object[]} All Bricks ids and titles
  */
 exports.getBricks = function () {
@@ -26,6 +40,11 @@ exports.getBricks = function () {
 exports.createBrick = function (brickObj) {
   brickObj.argScope = 0;
   brickObj.creationDate = new Date();
+
+  // check brick data
+  if (!brickObj.brief) { return new Promise((res, rej) => rej('Brief is required')); }
+  if (!brickObj.title) { return new Promise((res, rej) => rej('Title is required')); }
+
   return bricksRef.add(brickObj).then(ref => ref.id);
 }
 
@@ -35,8 +54,16 @@ exports.createBrick = function (brickObj) {
  * @param {object} brickObj object
  * @return {string} Brick Id
  */
-exports.updateBrick = function (brickId, brickObj) {
-  return brickRef(brickId).set(brickObj).then(ref => ref.id);
+exports.updateBrick = async function (brickId, brickObj) {
+  // check brick data
+  if (!brickObj.brief) { return new Promise((res, rej) => rej('Brief is required')); }
+  if (!brickObj.title) { return new Promise((res, rej) => rej('Title is required')); }
+
+  if (await getBrick(brickId)) {
+    return brickRef(brickId).set(brickObj).then(ref => ref.id);
+  } else {
+    return new Promise((res, rej) => rej('Brick with Id = ' + brickId + ' not found'));
+  }
 }
 
 /**
@@ -44,6 +71,10 @@ exports.updateBrick = function (brickId, brickObj) {
  * @param {string} brickId Brick Id
  * @return {string} Brick Id
  */
-exports.deleteBrick = function (brickId) {
-  return brickRef(brickId).delete().then(ref => ref.id);
+exports.deleteBrick = async function (brickId) {
+  if (await getBrick(brickId)) {
+    return brickRef(brickId).delete().then(ref => ref.id);
+  } else {
+    return new Promise((res, rej) => rej('Brick with Id = ' + brickId + ' not found'));
+  }
 }
